Solved 1. Two Sum
My thought process was to use a Dictionary (HashMap) to store the index-value pairs from the "nums" array. With the Dictionary, I could quickly look up the value needed to achieve the target sum, which would help reduce the time complexity.

I started by initializing the Dictionary and the "result" array to store the indices of the two numbers. Then, I iterated through the "nums" array and populated the Dictionary with the index-value pairs.

Next, I needed to find the first index of the pair that adds up to the target. I used a loop and removed elements from the Dictionary one by one, checking if the complementary value existed in the Dictionary's values. Once I found the first index, I broke out of the loop.

The last part of the code involved finding the second index. I used another loop, starting from the "first + 1" index, to check for the complementary value in the "nums" array. When I found it, I updated the "second" index and completed the solution.