Solved 58. Length of Last Word
Inside the class, there is a method named LengthOfLastWord, which takes a single argument s of type string. This method seems to be responsible for calculating the length of the last word in the input string.

The method begins by creating an array of strings named words. It does this by splitting the input string s using the Split() method without any specific delimiter. This means that it will split the string into substrings using spaces as the default delimiter. The Trim() method is called on the input string before splitting to remove any leading or trailing whitespace.

After splitting the input string into words, the code calculates the length of the last word. It does this by accessing the last element of the words array using the Length property. words.Length - 1 is used as the index to access the last element because arrays in C# are zero-indexed.

Finally, the method returns the length of the last word.