Solved 66. Plus One
Imagine you have a number like 123. We represent it as an array [1, 2, 3] where the first element (1) is the leftmost digit, the second element (2) is the next digit, and so on.

This code helps us add 1 to the number. Just like when you add 1 to 123, you get 124. But we're working with the array [1, 2, 3]. So, we start from the right end (the least significant digit), add 1 to it (like carrying over), and if it's bigger than 9, we carry over again.

We repeat this process, updating each digit and carrying over if needed. Once we finish this process, if there's still a carry left, we add it as an extra digit.

In the end, we reverse the order of digits because we've been adding digits from right to left, but we want the result to be in the proper order. Then, we convert our list of digits back to an array and return it.