Solved 121. Best Time to Buy and Sell Stock
The MaxProfit method takes an array of prices as input, where each element represents the price of the stock at a particular day.

It initializes two variables, minPrice and maxProfit, to keep track of the lowest price encountered and the maximum profit that can be obtained.

The code then loops through the prices array using a for loop.

    Inside the loop, it checks if the current stock price (at index i) is lower than the previously recorded minPrice. If it is, the minPrice is updated to the current price, ensuring that the variable always holds the lowest price encountered so far.

Next, the code calculates the potential profit that could be obtained by selling the stock at the current price (prices[i]) after buying it at the minPrice. If this potential profit is greater than the previously recorded maxProfit, the maxProfit is updated.

Once the loop has finished iterating through all the stock prices, the method returns the final maxProfit, representing the maximum profit that can be obtained by buying and selling the stock according to the historical price data.