Solved 274. H-Index
Finding the Highest Cited Number: The program starts by figuring out the highest number of citations any paper has received. This is like finding the scientist with the most citations among all the scientists.

Looping through Citations: The program enters a loop that does the following:

It sets a variable called "citedNumber" to 0. This variable will keep track of how many papers have been cited a certain number of times (we'll see how many times shortly).

It goes through each scientist's papers one by one (represented by the "for" loop).

If a scientist's paper has at least as many citations as the current number being checked (starting from the highest number), the "citedNumber" is increased by 1.

Checking H-Index Condition: After counting how many papers have citations equal to or greater than the current number, the program checks if the "citedNumber" is at least as large as the current number being checked. This is like asking if there are enough papers with that many citations. If there are, it means the H-Index has been found.

Decreasing the Checked Number: If the condition in step 3 isn't met, the program reduces the checked number by 1 and goes through the loop again. This is like saying, "Okay, let's check if there's an H-Index of one less."

Final Result: If the program finds a number where there are at least that many papers with that many citations, it means the H-Index is found. The program then returns that number. If it goes through all the numbers without finding an H-Index, it returns 0, indicating that no H-Index was found.