Solved 242. Valid Anagram
The first thing it does inside the method is to sort the characters in both strings 's' and 't.' Sorting is done in a specific way:

    Characters are first converted to uppercase to ensure case-insensitive comparison.
    Then, the characters are sorted in their natural order (ascending order, like alphabetical order).

After sorting, the 's' and 't' strings are updated to contain the sorted characters.

Next, it compares the sorted 's' and 't' strings. If they are equal, it means that the original 's' and 't' are anagrams of each other, so it returns true. Otherwise, it returns false.