Solved 1603. Design Parking System
The "AddCar" method is used to add a car of a specific type to the parking system. It takes the car type as an argument (1 for big, 2 for medium, 3 for small). If there are available parking spaces for that car type, the method will increment the count in the tracking system to reflect that a car has been parked, and it will return "true" to indicate success. If there are no available spaces for that car type, it returns "false," indicating that the car cannot be parked.