Solved 94. Binary Tree Inorder Traversal
    Input Validation: Check if the root of the tree is null. If it is, return an empty list as there's no tree to traverse.

    Create a Result List: Initialize an empty list called result to store the values obtained during the traversal.

    Traverse the Left Subtree: Perform an inorder traversal on the left subtree of the current node by recursively calling the InorderTraversal function on root.left. This step continues until it reaches the leftmost node of the tree.

    Add Current Node's Value: Add the value of the current node (root.val) to the result list. This happens when the recursive calls start to unwind back up the tree.

    Traverse the Right Subtree: Perform an inorder traversal on the right subtree of the current node by recursively calling the InorderTraversal function on root.right. This step continues until it reaches the rightmost node of the tree.

    Return the Result List: After completing the traversal, return the result list containing the values of the binary tree nodes in inorder traversal order.

Time Complexity: The time complexity of this algorithm is O(N), where N is the number of nodes in the binary tree. This is because the algorithm must visit each node exactly once.

Space Complexity: The space complexity of this algorithm is O(N), where N is the number of nodes in the binary tree. This space is used to store the values in the result list and also for the recursive call stack during traversal, which could potentially go as deep as the height of the tree.